Moved remaining model tests into tests/core
